[foilpics] Handle various decryption corner cases
The original directory may get removed after the file has been
encrypted. The most common case is a change or removal of the
SD-card. If the directory under home is removed, it gets recreated.
If a directory on SD-card no longer exists, the file is saved to
~/Pictures/FoilPics directory.

If the file with the same name already exists, it doesn't get
overwritten. Instead, a new file with a similar (albeit different)
name is created next to the already existing file.